requirements.txt
fastapi
uvicorn
pyTelegramBotAPI
python-dotenv
python-multipart
requests
